requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-admin-rangefilter==0.11.1
django-extensions==3.2.3
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
Pillow==10.0.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
yarl==1.9.2
